+++ mmm_eval/metrics/__init__.py
@@ -4,10 +4,10 @@ from .accuracy_functions import (
     calculate_absolute_percentage_change,
     calculate_mean_for_singular_values_across_cross_validation_folds,
     calculate_means_for_series_across_cross_validation_folds,
-    calculate_smape,
     calculate_std_for_singular_values_across_cross_validation_folds,
     calculate_stds_for_series_across_cross_validation_folds,
 )
+from .metric_models import calculate_smape
 
 __all__ = [
     "calculate_means_for_series_across_cross_validation_folds",
