+++ mmm_eval/metrics/metric_models.py
@@ -15,6 +15,41 @@ from mmm_eval.metrics.threshold_constants import (
 )
 
 
+def calculate_smape(actual: pd.Series, predicted: pd.Series) -> float:
+    """Calculate Symmetric Mean Absolute Percentage Error (SMAPE).
+    
+    SMAPE is calculated as: 100 * (2 * |actual - predicted|) / (|actual| + |predicted|)
+    
+    Args:
+        actual: Actual values
+        predicted: Predicted values
+        
+    Returns:
+        SMAPE value as float (percentage)
+        
+    Raises:
+        ValueError: If series are empty or have different lengths
+    """
+    # Validate inputs
+    if len(actual) == 0 or len(predicted) == 0:
+        raise ValueError("Cannot calculate SMAPE on empty series")
+    
+    if len(actual) != len(predicted):
+        raise ValueError("Actual and predicted series must have the same length")
+    
+    # Handle NaN values
+    if actual.isna().any() or predicted.isna().any():
+        return float(np.nan)
+    
+    # Handle division by zero and edge cases
+    denominator = np.abs(actual) + np.abs(predicted)
+    # Avoid division by zero by setting denominator to 1 where it's 0
+    denominator = np.where(denominator == 0, 1, denominator)
+    
+    smape = 100 * np.mean(2 * np.abs(actual - predicted) / denominator)
+    return float(smape)
+
+
 class MetricNamesBase(Enum):
     """Base class for metric name enums."""
 
@@ -188,15 +223,9 @@ class AccuracyMetricResults(MetricResults):
             AccuracyMetricResults object with the metrics
 
         """
-        # Calculate SMAPE inline to avoid circular import
-        denominator = np.abs(actual) + np.abs(predicted)
-        # Avoid division by zero by setting denominator to 1 where it's 0
-        denominator = np.where(denominator == 0, 1, denominator)
-        smape = 100 * np.mean(2 * np.abs(actual - predicted) / denominator)
-
         return cls(
-            mape=mean_absolute_percentage_error(actual, predicted),
-            smape=float(smape),
+            mape=mean_absolute_percentage_error(actual, predicted) * 100,
+            smape=calculate_smape(actual, predicted),
             r_squared=r2_score(actual, predicted),
         )
