+++ mmm_eval/metrics/threshold_constants.py
@@ -6,18 +6,18 @@
 class AccuracyThresholdConstants:
     """Constants for the accuracy threshold."""
 
-    MAPE = 0.15
-    SMAPE = 0.15
+    MAPE = 15.0
+    SMAPE = 15.0
     R_SQUARED = 0.8
 
 
 class CrossValidationThresholdConstants:
     """Constants for the cross-validation threshold."""
 
-    MEAN_MAPE = 0.15
-    STD_MAPE = 0.03
-    MEAN_SMAPE = 0.15
-    STD_SMAPE = 0.03
+    MEAN_MAPE = 15.0
+    STD_MAPE = 3.0
+    MEAN_SMAPE = 15.0
+    STD_SMAPE = 3.0
     MEAN_R_SQUARED = 0.8
