+++ tests/test_accuracy_functions.py
@@ -8,13 +8,13 @@ from mmm_eval.metrics.accuracy_functions import (
     calculate_absolute_percentage_change,
     calculate_mean_for_singular_values_across_cross_validation_folds,
     calculate_means_for_series_across_cross_validation_folds,
-    calculate_smape,
     calculate_std_for_singular_values_across_cross_validation_folds,
     calculate_stds_for_series_across_cross_validation_folds,
 )
 from mmm_eval.metrics.metric_models import (
     AccuracyMetricNames,
     AccuracyMetricResults,
+    calculate_smape,
 )
